Remove size_t

+++ src/test.cc
@@ -1,6 +1,4 @@
 
-#include <cstddef>
-
 namespace {
 
 namespace std {
@@ -14,8 +12,8 @@ namespace Kokkos {
 template <typename... A>
 class RangePolicy {
 public:
-  explicit RangePolicy(size_t in) {}
-  explicit RangePolicy(size_t in, size_t out) {}
+  explicit RangePolicy(int in) {}
+  explicit RangePolicy(int in, int out) {}
 };
 
 template <typename T>
@@ -46,20 +44,20 @@ void parallel_for_blocking(Policy rp) { }
 template <typename Policy, typename X>
 void parallel_for_blocking(Policy rp, X) { }
 
-void parallel_for_blocking(std::string str, size_t x) { }
-void parallel_for_blocking(size_t x) { }
-void parallel_scan_blocking(std::string str, size_t j) { }
-void parallel_scan_blocking(size_t i) { }
+void parallel_for_blocking(std::string str, int x) { }
+void parallel_for_blocking(int x) { }
+void parallel_scan_blocking(std::string str, int j) { }
+void parallel_scan_blocking(int i) { }
 
 template <typename Policy>
 void parallel_for_async(std::string str, Policy rp) { }
 template <typename Policy, typename X>
 void parallel_for_async(std::string str, Policy rp, X) { }
 
-void parallel_for_async(std::string str, size_t x) { }
-void parallel_for_async(size_t x) { }
-void parallel_scan_async(std::string str, size_t j) { }
-void parallel_scan_async(size_t i) { }
+void parallel_for_async(std::string str, int x) { }
+void parallel_for_async(int x) { }
+void parallel_scan_async(std::string str, int j) { }
+void parallel_scan_async(int i) { }
 
 void deep_copy_blocking(int, int);
 
@@ -97,12 +95,12 @@ void test() {
 #define KOKKOS_LAMBDA []
 
 struct X {
-  size_t extent(int len){return 0;};
+  int extent(int len){return 0;};
 };
 
 template <typename T>
 void test4() {
-  using Size = size_t;
+  using Size = int;
 
   Kokkos::parallel_for(20, []{});
   Kokkos::fence();
@@ -121,7 +119,7 @@ void test4() {
 
   using DeviceSpace = PHX::Device;
 
-  size_t N = 0;
+  int N = 0;
 
   Kokkos::parallel_for (Kokkos::RangePolicy<DeviceSpace,Kokkos::Schedule<Kokkos::Schedule<Kokkos::Dynamic>>>(0,N-2),
   []{});
